[step1] rename: util -> utils
